[changed] offset, limit types and date type of todo migration

+++ internal/ports/storage.go
@@ -25,7 +25,7 @@ type Storage interface {
 	Insert(ctx context.Context, data InsertDTO) error
 	Update(ctx context.Context, data UpdateDTO) error
 	Delete(ctx context.Context, id string) error
-	Get(ctx context.Context, offset int, limit int) ([]models.Todo, error)
+	Get(ctx context.Context, offset uint64, limit uint64) ([]models.Todo, error)
 	GetByDate(ctx context.Context, date time.Time) ([]models.Todo, error)
 	GetByID(ctx context.Context, id string) (*models.Todo, error)
 }
